pyproject.toml
[project]
name = "bcell"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-ollama>=0.2.2",
    "langchain>=0.3.15",
    "langchain-anthropic>=0.3.4",
    "python-dotenv>=1.0.1",
    "ollama>=0.4.7",
    "langgraph>=0.2.67",
    "transformers>=4.48.1",
    "tensorflow>=2.18.0",
    "langchain-pinecone>=0.2.2",
    "langchain-community>=0.3.15",
    "pypdf>=5.2.0",
    "pinecone-client>=4.1.2",
    "langchain-openai>=0.3.2",
    "langchain-experimental>=0.3.4",
    "unstructured>=0.16.17",
    "langchain-neo4j>=0.3.0",
    "elevenlabs>=1.50.7",
    "langchain-google-genai>=2.0.9",
    "fastapi[all]>=0.115.8",
]
